+++ smath/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 1.4.0
+
+- Add `error()` function
+- Update `exray` dependency version
+- Add documentation on `npx smath` executables
+
 ## 1.3.5
 
 - Restructure test files and scripts